+++ test/main.cpp
@@ -9,10 +9,10 @@ int main()
         int aa = 10;
         std::string bb = "Hello";
 
-        logger.debug("This is a debug message");
-        logger.info("This is an info message");
-        logger.warning("This is a warning message");
-        logger.error("This is an error message");
+        logger.debug("This is a debug message ", aa);
+        logger.info("This is an info message ", aa, bb);
+        logger.warning("This is a warning message ","test", aa, bb);
+        logger.error("This is an error message ", "test---", " ", aa, bb);
 
     } catch (const std::exception &e) {
         std::cerr << "Exception: " << e.what() << std::endl;
